Fixed the bigint implementation.
The bigint would not properly handle all sorts of special cases. Now there is
a 100% coverage test, at least. There may still be a few edges cases that are
not checked properly, but it's much less likely.

Actually make use of the add.s and added the sub.s file now.

Fixed the function names so they are C++ (and not C function names).

Removed snapdatabase_... from the exception names.

Fixed bug where a memset() was used with "f_value + pos * 8", when it
should have been "f_value + pos". The compiler does the "*8" automatically.

Make use of the zero() and is_zero() functions instead of loops or memset().

Fixed the div() function. Many mistakes in there!

Updated the TODO since the XML support was moved to a separate project.