Added a test to verify the hashing.
There was actually a bug in the streaming feature where the get_byte() was
used in the get() function when it was supposed to be peek_byte().

Made the bigint test cover 100% of the code and fixed many problems.

Fixed the bloom filter hash function. I commented out something which should
not have been commented.