Started work on a journal tool.
Got the journal compression code to work. Since this is expected to be a
relatively small set of files, the compression is done all at once in a
file as we load it (the load also writes). I also added a magic code to
the start of each event so we can know whether we reached the end of the
list or not (i.e. the compression does not truncate the file by default).