Fixed memory leak.
I made the vector of files in the journal class a vector of weak pointers.
So that way when all the events in a file as held by the location lists
goes empty, the corresponding file properly gets deleted in all circumstances.

It also means we do more open/close of files since we do not keep an open
version in memory if the file is empty (even if it has a header).