Started work on having a prinbee SQL-like language
This version is heavily tweaking the settings to skip on XML files which
are rather hard to update, compare, etc. Instead, prinbee will use simple
.ini files.

Further, I looked into better handling of the language and versioning of
the schemata of a context. Renamed many get_...() functions. Removed things
we are not going to use anymore (sparse flag).

And many other things.