Finished implementing the language-manager tool.
This loads the languages as defined by ICU and converts that to our own
format with an identifier we can use in our Prinbee database. The
identifier is just 16 bits that are used to distinguish between versions
of a row (i.e. one row [a.k.a. one key] can then be in German and French).
Your application can specify a language and a default language and the
GET from the database always returns some data, as close as possible to
what was requested.