Changed context to use a setup.
Before I used an advgetopt object, which works too, but it implies we
have a .conf and additional parameters in that .conf when now I am thinking
all the data should be within the context path. So the only two things we
need now is the path and ownership. Both have a default specific to prinbee
(i.e. assuming most people will want to run one single context per computer).