Includes many more node tests for nearly 100% coverage.
The parser now generates a tree of nodes, then calls a function to output
the as2js code. This is cleaner and a bit easier to test the validity.
I am still missing a lot of parenthesis (i.e. order of operations is
completely different in SQL than regular C-like languages like JavaScript).