Moved the function as postfix.
I tested further thinking that functions just could not be in the "other"
group. I moved them to postfix and now that works as expected. This means
that the casting `<type> [(] <expr> [)]` also had to move there. This
also means we have many different ways to cast and I think it's ugly but
that matches (more or less) what Postgres supports.