EdHttp does not seem necessary. Removing for now.
I had an error which was probably because I did not recompile the whole
stack. Removing this dependency still get things compiled now. That
way I avoid the dependency in the control file. That being said, at
some point we probably will want to directly use edhttp and it may also
fail in the build system.