Started to get the HELP; command working.
Also I fixed a bug with the ping timer which may ping us before
the binary connection is ready. I am thinking to move the code
that creates that timer AFTER we get the binary connection so
that way we avoid wasting time/memory (other resources?) if
there isn't a connection.