Started fixing the ipwall / iplock status.
The firewall status is now managed in a communicatord plugin. We do not
have to create a wait_on_firewall object (which we could not do from
prinbee anyway since iplock depends on prinbee...)

So now we send an IPLOCK_GET_STATUS message and accept IPLOCK_CURRENT_STATUS
messages from the communicatord. Also we properly check the status as either
UP or ACTIVE.